+++ lib/outpost/list.rb
@@ -4,9 +4,9 @@ module Outpost
   module List
     extend ActiveSupport::Autoload
 
-    DEFAULT_ORDER     = "id"
-    DEFAULT_SORT_MODE = "desc"
-    DEFAULT_PER_PAGE  = 25
+    DEFAULT_ORDER_ATTRIBUTE   = "id"
+    DEFAULT_ORDER_DIRECTION   = "DESC"
+    DEFAULT_PER_PAGE          = 25
 
     autoload :Base
     autoload :Column