+++ spec/lib/list/base_spec.rb
@@ -10,36 +10,36 @@ describe Outpost::List::Base do
     context "with block" do
       it "yields and sets attributes" do
         list = Outpost::List::Base.new model do |l|
-          l.default_order     = "omg"
-          l.default_sort_mode = "wtf"
-          l.per_page          = 999
+          l.default_order_attribute   = "omg"
+          l.default_order_direction   = "wtf"
+          l.per_page                  = 999
         end
 
-        list.default_order.should     eq "omg"
-        list.default_sort_mode.should eq "wtf"
-        list.per_page.should          eq 999
+        list.default_order_attribute.should eq "omg"
+        list.default_order_direction.should eq "wtf"
+        list.per_page.should eq 999
       end
 
       it "sets any attributes that weren't set in the block" do
         list = Outpost::List::Base.new model do |l|
-          l.default_order     = "omg"
+          l.default_order_attribute = "omg"
         end
 
-        list.default_order.should     eq "omg"
-        list.default_sort_mode.should eq Outpost::List::DEFAULT_SORT_MODE
-        list.per_page.should          eq Outpost::List::DEFAULT_PER_PAGE
+        list.default_order_attribute.should eq "omg"
+        list.default_order_direction.should eq Outpost::List::DEFAULT_ORDER_DIRECTION
+        list.per_page.should eq Outpost::List::DEFAULT_PER_PAGE
       end
     end
 
     context "without block" do
       it "sets order to default" do
         list = Outpost::List::Base.new(model)
-        list.default_order.should eq Outpost::List::DEFAULT_ORDER
+        list.default_order_attribute.should eq Outpost::List::DEFAULT_ORDER_ATTRIBUTE
       end
 
-      it "sets sort_mode to default" do
+      it "sets order_direction to default" do
         list = Outpost::List::Base.new(model)
-        list.default_sort_mode.should eq Outpost::List::DEFAULT_SORT_MODE
+        list.default_order_direction.should eq Outpost::List::DEFAULT_ORDER_DIRECTION
       end
 
       it "sets per_page to default" do